# docs/swagger.yaml
basePath: /api
definitions:
...
  domain.WarehouseCreateRequest:
    properties:
      is_active:
        example: true
        type: boolean
      name:
        example: Main Warehouse
        type: string
      shop_id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
    required:
    - name
    - shop_id
    type: object
host: localhost:8080
info:
  contact: {}
  description: This is a comprehensive warehouse management system API for handling
    products, orders, warehouses, and transfers.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Warehouse Management API
  version: "1.0"
paths:
...
  /shops/{shop_id}/warehouses:
    get:
      consumes:
      - application/json
      description: Retrieve all warehouses associated with a specific shop
      parameters:
      - description: Shop ID (UUID)
        format: uuid
        in: path
        name: shop_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Warehouses retrieved successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid shop ID format
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get warehouses by shop ID
      tags:
      - Warehouses
...
  /warehouses:
    post:
      consumes:
      - application/json
      description: Create a new warehouse associated with a shop
      parameters:
      - description: Warehouse creation data
        in: body
        name: warehouse
        required: true
        schema:
          $ref: '#/definitions/domain.WarehouseCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Warehouse created successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request payload or validation failed
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create a new warehouse
      tags:
      - Warehouses
  /warehouses/{id}:
    delete:
      consumes:
      - application/json
      description: Delete an existing warehouse by its UUID
      parameters:
      - description: Warehouse ID (UUID)
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Warehouse deleted successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid warehouse ID format
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Warehouse not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete warehouse
      tags:
      - Warehouses
    get:
      consumes:
      - application/json
      description: Retrieve a specific warehouse by its UUID
      parameters:
      - description: Warehouse ID (UUID)
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Warehouse retrieved successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid warehouse ID format
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Warehouse not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get warehouse by ID
      tags:
      - Warehouses
    put:
      consumes:
      - application/json
      description: Update an existing warehouse by its UUID
      parameters:
      - description: Warehouse ID (UUID)
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Updated warehouse data
        in: body
        name: warehouse
        required: true
        schema:
          $ref: '#/definitions/domain.WarehouseCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Warehouse updated successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid warehouse ID or request payload
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Warehouse not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update warehouse
      tags:
      - Warehouses
  /warehouses/{id}/status:
    patch:
      consumes:
      - application/json
      description: Enable or disable a warehouse by updating its active status
      parameters:
      - description: Warehouse ID (UUID)
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Active status data
        in: body
        name: status
        required: true
        schema:
          properties:
            is_active:
              type: boolean
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Warehouse status updated successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid warehouse ID or request payload
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Warehouse not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Set warehouse active status
      tags:
      - Warehouses
...
produces:
- application/json
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"